Replace high/low accuracy with omit rests option.
I have been feeling that the high/low accuracy switch is a bit
misleading because while it can help us choose waypoints that
are sufficiently far apart from each other, the accuracy of any
individual waypoint isn't inherently better than normal. However,
the option to omit or keep rests allows the user to keep the raw
data or postprocess it in their own way; or accept the default
different-enough detection.