Refactor out TracklistElement, make Track more object-oriented.
I'm sure the reason for TracklistElement was to save memory by not
having the waypoints loaded, but the separation of Track /
TracklistElement creates lots of back-and-forth conversions and
mental overhead. I want to give this a try and see if it actually
causes any problems.

Moved a lot of helper functions into the classes they operate on.