Remove coroutine when deleting track, just do it synchronously.
My coroutine solution worked for bringing the onboarding back, but
it broke the cancel button since it touched a view out of the owning
thread. I probably could fix that by adding a coroutine for cancel
too, but the delete operation should be so fast that synchronous
might not even be a problem.