Add wakelock to prevent doze when not at home.
Recently I took a 90 minute drive, and trkpt missed about 15
minutes of it. It wasn't until now that I realized it probably
went into doze mode due to the low accelerometer movements.
If the device is plugged in it does not doze, but to achieve the
no-friction goal we should not be required to take it out of our
pocket and plug in all the time for fear of missing points.