Reload the database when resuming full_power.
I had an incident where I was tampering with the database on my
computer, being synced over Syncthing, and I caused a conflict that
made the app stop recording, even though the app UI continued to show
the normal recording icons. Closing and re-opening the db connection
when leaving the house would have helped prevent that.