Fix behavior of track delete button after using when was I here.
The delete button was deleting all points between the first and last,
which, when you use the "when was I here" button, could span days
or months that you don't want to delete. So, after doing a spatial
query, the delete button will only delete the points that are shown.